Fix majority of compiler warnings
Also disabled the struct warning when assigning to {0}. This is legal as per the C99 spec, but clang freaks out about it.

+++ src/dictionary/bpp_tree/bpp_tree.h
@@ -66,7 +66,7 @@ typedef struct {
 	char		*iName;			/* name of index file */
 	int			keySize;		/* length, in bytes, of key */
 	bpp_bool_t	dupKeys;				/* true if duplicate keys allowed */
-	int			sectorSize;		/* size of sector on disk */
+	size_t		sectorSize;		/* size of sector on disk */
 	bCompType	comp;			/* pointer to compare function */
 } bOpenType;
